requirements.txt
alembic==1.10.4
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
protobuf==3.20.3
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
pytest==7.3.1
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
